requirements.txt
fastapi
uvicorn
python-dotenv
httpx
markdown
beautifulsoup4
python-docx
PyPDF2
py-zerox
google-generativeai
anthropic
openai
